chatbot/requirements.txt
google-generativeai
gradio
python-dotenv
rich
tiktoken
socksio